Maintenance
- phpstan \o/

+++ src/Year2022/Day03/PartOne.php
@@ -11,7 +11,7 @@ use function Safe\preg_match_all;
 class PartOne
 {
     /**
-     * @param Generator<void, string, void, void> $input
+     * @param Generator<string> $input
      *
      * @return int
      */
@@ -22,7 +22,7 @@ class PartOne
         $sum = 0;
 
         foreach ($input as $line) {
-            $l2 = strlen($line) / 2;
+            $l2 = (int)(strlen($line) / 2);
             $first = substr($line, 0, $l2);
             $second = substr($line, $l2);
